Decouple core settings from plugin middleware.
