+++ kolibri/core/device/middleware.py
@@ -1,8 +1,13 @@
+from django.core.urlresolvers import reverse
 from django.http import HttpResponse
 from django.middleware.locale import LocaleMiddleware
+from django.shortcuts import redirect
 from django.utils import translation
+from django.utils.deprecation import MiddlewareMixin
 
 from .translation import get_language_from_request
+from kolibri.core.device.hooks import SetupHook
+from kolibri.core.device.utils import device_provisioned
 
 
 class KolibriLocaleMiddleware(LocaleMiddleware):
@@ -33,3 +38,36 @@ class IgnoreGUIMiddleware(object):
         if request.META.get("HTTP_USER_AGENT", None) == "Kolibri session":
             return HttpResponse('')
         return None
+
+
+ALLOWED_PATH_LIST = [
+    "kolibri:deviceprovision",
+    "kolibri:setupwizardplugin:setupwizard",
+    "kolibri:set_language",
+    "kolibri:session-list"
+]
+
+SETUP_WIZARD_URLS = [hook.url for hook in SetupHook().registered_hooks]
+
+
+class SetupMiddleware(MiddlewareMixin):
+    """
+    display the setup wizard if device is not provisioned
+    """
+    device_provisioned = False
+
+    def process_request(self, request):
+        # If a DevicePermissions with is_superuser has already been created, no need to do anything here
+        self.device_provisioned = self.device_provisioned or device_provisioned()
+        if self.device_provisioned:
+            if any(map(lambda x: request.path.startswith(x), SETUP_WIZARD_URLS)):
+                return redirect(reverse("kolibri:redirect_user"))
+            return
+
+        # Don't redirect for URLs that are required for the setup wizard
+        allowed_paths = [reverse(name) for name in ALLOWED_PATH_LIST]
+        if any(request.path.startswith(path_prefix) for path_prefix in allowed_paths):
+            return
+        if SETUP_WIZARD_URLS:
+            # If we've gotten this far, we want to redirect to the setup wizard
+            return redirect(SETUP_WIZARD_URLS[0])
