+++ kolibri/deployment/default/settings/base.py
@@ -90,7 +90,7 @@ MIDDLEWARE = [
     'kolibri.core.device.middleware.KolibriLocaleMiddleware',
     'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
-    'kolibri.plugins.setup_wizard.middleware.SetupWizardMiddleware',
+    'kolibri.core.device.middleware.SetupMiddleware',
     'kolibri.core.auth.middleware.CustomAuthenticationMiddleware',
     'kolibri.core.middleware.signin_page.RedirectToSignInPageIfNoGuestAccessAndNoActiveSession',
     'django.contrib.auth.middleware.SessionAuthenticationMiddleware',
