+++ kolibri/plugins/setup_wizard/kolibri_plugin.py
@@ -3,6 +3,7 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 from . import hooks
+from kolibri.core.device.hooks import SetupHook
 from kolibri.core.webpack import hooks as webpack_hooks
 from kolibri.plugins.base import KolibriPluginBase
 
@@ -23,3 +24,9 @@ class SetupWizardAsset(webpack_hooks.WebpackBundleHook):
 
 class SetupWizardInclusionHook(hooks.SetupWizardSyncHook):
     bundle_class = SetupWizardAsset
+
+
+class SetupWizardHook(SetupHook):
+    @property
+    def url(self):
+        return self.plugin_url(SetupWizardPlugin, 'setupwizard')
